re-factor event-listeners
